# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spool API",
    "version": "1.0.0",
    "description": "The REST API of a Spool server: a self-hosted git forge that keeps every repository in object storage. Git smart-HTTP endpoints (clone/fetch/push at /{org}/{repo}.git) speak the standard git wire protocol and are not described here. Every repository is private to the organization that owns it: a request with no credential answers 401 on every repository, changeset, search and profile route, and a credential that may not read a repository gets the same 404 a repository that does not exist gets. The exceptions are the sign-in routes under /v1/auth, the health and metrics probes, the inbound webhook receiver, the runner routes (which take their own credentials) and the signed CDN pack URL. Errors return JSON {\"error\": \"…\"} with the matching status. Optimistic-concurrency conflicts return 409 with the current tip."
  },
  "servers": [
    {
      "url": "https://spool.example.com",
      "description": "Your server; set by STRATUM_PUBLIC_URL."
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token of the form weft_<id>_<secret>. Scopes: org:admin, org:read, repo:read, repo:write. A token acts only in the organization it was minted in: presented anywhere else it has no access, and a repository there answers 404. Repo-bound tokens are restricted to one repository and cannot search."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "stratum_session",
        "description": "Browser session from POST /v1/auth/login. An alternative to a bearer token for the dashboard; a bearer token wins when both are presented. A session spans every organization its person belongs to, which is why anything that crosses two namespaces — forking, opening a change from a fork — takes one. Routes under /v1/auth use this scheme."
      }
    },
...
  },
  "paths": {
...
    "/v1/github/setup": {
      "get": {
        "summary": "GitHub sends a browser here after an App install or an edit to one. Two proofs, for two questions. Whose installation: the `code` GitHub appends when the App requests user authorization during installation is exchanged for that person's token, and the installation must be among the ones `GET /user/installations` lists for them — otherwise `connect=notyours` and nothing is bound. Which organization: the single-use `state` minted by `POST /v1/orgs/{org}/github/install`; without one (GitHub sends none after an edit, and an install begun on GitHub or from the Marketplace never had one) the signed-in person's own single pending connect stands in; otherwise the installation, already proved theirs, is parked and the browser is sent to `connect=claim` with a claim cookie, and `POST /v1/orgs/{org}/github/install/claim` binds it after sign-in. Answers a redirect back into the dashboard either way, carrying `connect=ok|claim|notyours|missing|expired|taken|error` — every dead state is `expired`, so a callback cannot be used to learn which flows exist.",
        "parameters": [
          {
            "name": "installation_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The single-use state from `POST /v1/orgs/{org}/github/install`. GitHub sends none after an installation is edited; the signed-in person's own pending connect then names the organization."
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "GitHub's user-authorization code, appended when the App requests user authorization during installation. Exchanged for the person's token to check the installation is theirs; required when the deployment has an OAuth client configured."
          }
        ],
        "responses": {
          "303": {
            "description": "redirect to the dashboard with the outcome in `connect`"
          }
        }
      }
    },
...
  }
}